% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile}
\alias{compile}
\title{Build a function for transforming parameters into a valid path}
\usage{
compile(path, delimiter = "/", encode)
}
\description{
The function will have one parameter in which you can specify the parameters
}
