% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse}
\alias{parse}
\title{Decompose a pathname}
\usage{
parse(path)
}
\arguments{
\item{path}{A character vector of length 1.}
}
\value{
A list with the pathname's tokens.
}
\description{
Decompose a pathname
}
\details{
This function decomposes a given pathname into tokens.
}
