% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringify.R
\name{stringifyTokens}
\alias{stringifyTokens}
\title{From \code{tokenData} object to a character vector}
\usage{
stringifyTokens(tokens)
}
\arguments{
\item{tokens}{An object of class \code{tokenData}}
}
\value{
A character vector of length 1.
}
\description{
The inverse of the \code{\link[=parse]{parse()}} function.
}
\examples{
tokens <- parse("/path/to/resource/:Id")
path <- stringifyTokens(tokens)
identical(path, "/path/to/resource/:Id") # TRUE

}
