% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse}
\alias{parse}
\title{Decompose a pathname}
\usage{
parse(path, encodePath = utils::URLencode)
}
\arguments{
\item{path}{A character vector of length 1.}

\item{encodePath}{A function to encode characters, defaults to \code{URLencode}.}
}
\value{
An object of class \code{tokenData}.
}
\description{
This function decomposes a given pathname into meaningful tokens.
}
\details{
TODO
}
